backend/requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
python-jose==3.3.0
python-keycloak==1.3.0
requests==2.31.0
